+++ react/src/components/Breadcrumb/index.js
@@ -1,0 +1,1 @@
+export { Breadcrumb, BreadcrumbItem, BreadcrumbSeparator, BreadcrumbCurrent, ChevronThinIcon, ChevronThickIcon, default } from './Breadcrumb';